add offset limit query for getallbook
update response

+++ controllers/bookController.js
@@ -2,15 +2,37 @@ const { PrismaClient } = require('@prisma/client');
 const prisma = new PrismaClient();
 
 const getBooks = async (req, res) => {
+  const limit = parseInt(req.query.limit) || 6;
+  const offset = parseInt(req.query.offset) || 0
+  const search = req.query.search || "";
+  const category = req.query.category || "";
   try {
     const books = await prisma.book.findMany({
-      include: {
-        categories: true,
+      where: {
+        name: {
+          contains: search
+        },
+        categories: {
+          some: {
+            category: {
+              name: {
+                contains: category
+              }
+            }
+          }
+        }
       },
+      skip: offset,
+      take: limit,
     });
-    res.json(books);
+
+    res.status(200).json({
+      message: "success get data",
+      data: books
+    });
+
   } catch (error) {
-    res.status(500).json({ error: 'Error fetching books' });
+    res.status(500).json({ error: error.message});
   }
 };
 
@@ -36,10 +58,13 @@ const createBook = async (req, res) => {
         },
       },
     });
-    res.json(book);
+    res.status(201).json({
+      message: "create success",
+      data: book
+    });
   } catch (error) {
     console.error('Error creating book:', error);
-    res.status(500).json({ error: 'Error creating book' });
+    res.status(500).json({ error: error.message });
   }
 };
 
@@ -48,6 +73,9 @@ const updateBook = async (req, res) => {
   const { name, description, stock, imageurl, categoryIds } = req.body;
 
   try {
+
+    await checkBook(parseInt(id));
+
     const updatedData = {
       name,
       description,
@@ -55,27 +83,41 @@ const updateBook = async (req, res) => {
       imageurl,
     };
 
-    // Jika categoryIds ada dan merupakan array, proses kategori
-    if (Array.isArray(categoryIds) && categoryIds.length > 0) {
-      // Hapus kategori yang lama terlebih dahulu
-      await prisma.bookCategories.deleteMany({
-        where: { bookId: parseInt(id) },
-      });
+    await prisma.$transaction(async (prisma) => {
+      if (Array.isArray(categoryIds) && categoryIds.length > 0) {
+        // Delete old categories
+        await prisma.bookCategories.deleteMany({
+          where: { bookId: parseInt(id) },
+        });
 
-      // Tambahkan kategori yang baru
-      updatedData.categories = {
-        create: categoryIds.map(categoryId => ({
-          category: { connect: { id: categoryId } },
-        })),
-      };
+        // Add new categories
+        await prisma.bookCategories.createMany({
+          data: categoryIds.map(categoryId => ({
+            bookId: parseInt(id),
+            categoryId: categoryId,
+          })),
+        });
+      }
+    });
+
+    if(await checkBook(parseInt(id))){
+      return res.status(404).json({
+        message: "book not found"
+      })
     }
 
     const book = await prisma.book.update({
       where: { id: parseInt(id) },
       data: updatedData,
+      include: {
+        categories: true
+      }
     });
 
-    res.json(book);
+    return res.status(200).json({
+      message: "success update data",
+      data: book
+    });
   } catch (error) {
     console.error('Error updating book:', error);
     res.status(500).json({ error: 'Error updating book' });
@@ -87,6 +129,13 @@ const deleteBook = async (req, res) => {
   const { id } = req.params;
 
   try {
+
+    if(await checkBook(parseInt(id))){
+      return res.status(404).json({
+        message: "book not found"
+      })
+    }
+
     // Hapus relasi terkait terlebih dahulu (BookCategories, Peminjaman, CommentBook)
     await prisma.bookCategories.deleteMany({
       where: { bookId: parseInt(id) },
@@ -97,12 +146,20 @@ const deleteBook = async (req, res) => {
       where: { id: parseInt(id) },
     });
 
-    res.json({ message: 'Book deleted successfully' });
+    return res.status(200).json({ message: 'Book deleted successfully' });
   } catch (error) {
     console.error('Error deleting book:', error);
     res.status(500).json({ error: 'Error deleting book' });
   }
 };
+
+async function checkBook(id) {
+  return !await prisma.book.findUnique({
+    where: {
+      id
+    }
+  });
+}
 
 
 module.exports = {
